ferredesk_v0/backend/requirements.txt
#Se cambió Pillow a 11.2.1 para evitar errores de instalación en Windows
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==11.2.1
django-filter==23.5
dj-database-url==2.1.0
openpyxl==3.1.5
django-extensions==3.2.3
pyexcel==0.7.0
reportlab==4.4.2 
cryptography==45.0.5
requests==2.32.4
lxml==6.0.0
zeep==4.3.1
python-dateutil==2.9.0
